cuong update giaodien trang chu

+++ src/main/java/org/example/sd_94vs1/entity/oder/OrderLine.java
@@ -5,8 +5,6 @@ import lombok.*;
 import lombok.experimental.FieldDefaults;
 import org.example.sd_94vs1.entity.product.Product;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -31,7 +29,7 @@ public class OrderLine {
     Integer quantity;
 
     @Column(name = "price_at_order", nullable = false)
-    BigDecimal priceAtOrder;
+    Double priceAtOrder;
 
     @ManyToOne
     @JoinColumn(name = "order_code", referencedColumnName = "order_code", insertable = false, updatable = false)
